:art: style: 代码优化与 clippy 修复
- runtime/engine: 简化 dedup 和 contains 调用
- formatter: 移除冗余变量
- dead_code: 使用 cloned 替代手动 clone
- codegen: 简化 file_id 转换
- gen_interfaces: 使用 next_back 替代 last
- diagnostic: 移除冗余引用

+++ src/middle/passes/codegen/bytecode.rs
@@ -106,7 +106,7 @@ impl DebugSection {
             out.write_all(&(entries.len() as u32).to_le_bytes())?;
             for (ip, ds) in entries {
                 out.write_all(&(ip as u32).to_le_bytes())?;
-                out.write_all(&(ds.file_id as u32).to_le_bytes())?;
+                out.write_all(&ds.file_id.to_le_bytes())?;
                 write_position(&mut out, ds.span.start)?;
                 write_position(&mut out, ds.span.end)?;
             }
